Help HDF find libsz on non-linux architectures.
Newer versions of hdf need libsz.  Makefile.PL was hardcoded to look for libsz.so.  I added .dylib and .dll suffixes for OSX and mswin32 respectively. The default remains to look for .so, so we're no worse off than before. Probably there is a better way to do that, though.